+++ CHANGELOG.md
@@ -15,12 +15,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.4.1] - 2024-??-??
+
+### Added
+
+- example for usage of components (needs at least dimas 0.4.2)
+
+### Changed
+
+- names of multi session examples
+
+### Fixed
+
+### Removed
+
 ## [0.4.0] - 2024-11-13
 
 ### Added
 
 - example for multi session agents (needs at least dimas v0.4.1)
-- example for usage of components (needs at least dimas 0.4.2)
 
 ### Changed